// client/src/generated/paytowin.ts
export interface PlayersRequest {
}
/**
 * @generated from protobuf message PayToWin.PlayersResponse
 */
export interface PlayersResponse {
    /**
     * @generated from protobuf field: repeated string players = 1;
     */
    players: string[];
}
/**
 * @generated from protobuf message PayToWin.PotionEffect
 */
export interface PotionEffect {
    /**
     * @generated from protobuf field: PayToWin.PotionName name = 1;
     */
    name: PotionName;
    /**
     * @generated from protobuf field: uint32 duration = 2;
     */
    duration: number;
    /**
     * @generated from protobuf field: uint32 amplifier = 3;
     */
    amplifier: number;
}
/**
 * @generated from protobuf message PayToWin.Item
 */
export interface Item {
    /**
     * @generated from protobuf field: PayToWin.MinecraftMaterial itemName = 1;
     */
    itemName: MinecraftMaterial;
    /**
     * @generated from protobuf field: uint32 amount = 2;
     */
    amount: number;
}
/**
 * @generated from protobuf message PayToWin.SpawnEntity
 */
export interface SpawnEntity {
    /**
     * @generated from protobuf field: PayToWin.MinecraftEntity entity = 1;
     */
    entity: MinecraftEntity;
    /**
     * @generated from protobuf field: uint32 amount = 2;
     */
    amount: number;
}
/**
 * @generated from protobuf message PayToWin.EffectRequest
 */
export interface EffectRequest {
    /**
     * @generated from protobuf oneof: effect
     */
    effect: {
        oneofKind: "potion";
        /**
         * @generated from protobuf field: PayToWin.PotionEffect potion = 2;
         */
        potion: PotionEffect;
    } | {
        oneofKind: "dataless";
        /**
         * @generated from protobuf field: PayToWin.DatalessEffect dataless = 3;
         */
        dataless: DatalessEffect;
    } | {
        oneofKind: "tool";
        /**
         * @generated from protobuf field: PayToWin.DiamondTool tool = 4;
         */
        tool: DiamondTool;
    } | {
        oneofKind: "item";
        /**
         * @generated from protobuf field: PayToWin.Item item = 5;
         */
        item: Item;
    } | {
        oneofKind: "spawnEntity";
        /**
         * @generated from protobuf field: PayToWin.SpawnEntity spawnEntity = 6;
         */
        spawnEntity: SpawnEntity;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf field: string player = 1;
     */
    player: string;
    /**
     * @generated from protobuf field: string requester = 7;
     */
    requester: string;
}
/**
 * @generated from protobuf message PayToWin.EffectResponse
 */
export interface EffectResponse {
    /**
     * @generated from protobuf field: bool success = 1;
     */
    success: boolean;
}
/**
 * @generated from protobuf enum PayToWin.PotionName
 */
export enum PotionName {
    /**
     * @generated from protobuf enum value: SPEED = 0;
     */
    SPEED = 0,
    /**
     * @generated from protobuf enum value: SLOW = 1;
     */
    SLOW = 1,
    /**
     * @generated from protobuf enum value: FAST_DIGGING = 2;
     */
    FAST_DIGGING = 2,
    /**
     * @generated from protobuf enum value: SLOW_DIGGING = 3;
     */
    SLOW_DIGGING = 3,
    /**
     * @generated from protobuf enum value: INCREASE_DAMAGE = 4;
     */
    INCREASE_DAMAGE = 4,
    /**
     * @generated from protobuf enum value: HEAL = 5;
     */
    HEAL = 5,
    /**
     * @generated from protobuf enum value: HARM = 6;
     */
    HARM = 6,
    /**
     * @generated from protobuf enum value: JUMP = 7;
     */
    JUMP = 7,
    /**
     * @generated from protobuf enum value: CONFUSION = 8;
     */
    CONFUSION = 8,
    /**
     * @generated from protobuf enum value: REGENERATION = 9;
     */
    REGENERATION = 9,
    /**
     * @generated from protobuf enum value: DAMAGE_RESISTANCE = 10;
     */
    DAMAGE_RESISTANCE = 10,
    /**
     * @generated from protobuf enum value: FIRE_RESISTANCE = 11;
     */
    FIRE_RESISTANCE = 11,
    /**
     *  WATER_BREATHING = 12;
     *
     * @generated from protobuf enum value: INVISIBILITY = 13;
     */
    INVISIBILITY = 13,
    /**
     * @generated from protobuf enum value: BLINDNESS = 14;
     */
    BLINDNESS = 14,
    /**
     * @generated from protobuf enum value: NIGHT_VISION = 15;
     */
    NIGHT_VISION = 15,
    /**
     * @generated from protobuf enum value: HUNGER = 16;
     */
    HUNGER = 16,
    /**
     * @generated from protobuf enum value: WEAKNESS = 17;
     */
    WEAKNESS = 17,
    /**
     * @generated from protobuf enum value: POISON = 18;
     */
    POISON = 18,
    /**
     * @generated from protobuf enum value: WITHER = 19;
     */
    WITHER = 19,
    /**
     * @generated from protobuf enum value: HEALTH_BOOST = 20;
     */
    HEALTH_BOOST = 20,
    /**
     * @generated from protobuf enum value: ABSORPTION = 21;
     */
    ABSORPTION = 21,
    /**
     *  SATURATION = 22;
     *
     * @generated from protobuf enum value: GLOWING = 23;
     */
    GLOWING = 23,
    /**
     * @generated from protobuf enum value: LEVITATION = 24;
     */
    LEVITATION = 24,
    /**
     *  LUCK = 25;
     *  UNLUCK = 26;
     *
     * @generated from protobuf enum value: SLOW_FALLING = 27;
     */
    SLOW_FALLING = 27,
    /**
     *  CONDUIT_POWER = 28;
     *  DOLPHINS_GRACE = 29;
     *  BAD_OMEN = 30;
     *  HERO_OF_THE_VILLAGE = 31;
     *
     * @generated from protobuf enum value: DARKNESS = 32;
     */
    DARKNESS = 32
}
/**
 * @generated from protobuf enum PayToWin.DiamondTool
 */
export enum DiamondTool {
    /**
     * @generated from protobuf enum value: Boots = 0;
     */
    Boots = 0,
    /**
     * @generated from protobuf enum value: Leggings = 1;
     */
    Leggings = 1,
    /**
     * @generated from protobuf enum value: Chestplate = 2;
     */
    Chestplate = 2,
    /**
     * @generated from protobuf enum value: Helmet = 3;
     */
    Helmet = 3,
    /**
     * @generated from protobuf enum value: Axe = 4;
     */
    Axe = 4,
    /**
     * @generated from protobuf enum value: Hoe = 5;
     */
    Hoe = 5,
    /**
     * @generated from protobuf enum value: Pickaxe = 6;
     */
    Pickaxe = 6,
    /**
     * @generated from protobuf enum value: Shovel = 7;
     */
    Shovel = 7,
    /**
     * @generated from protobuf enum value: Sword = 8;
     */
    Sword = 8
}
/**
 * @generated from protobuf enum PayToWin.DatalessEffect
 */
export enum DatalessEffect {
    /**
     * @generated from protobuf enum value: RandomPotions = 0;
     */
    RandomPotions = 0,
    /**
     * @generated from protobuf enum value: Spinny = 1;
     */
    Spinny = 1,
    /**
     * @generated from protobuf enum value: PutInAdventure = 2;
     */
    PutInAdventure = 2,
    /**
     * @generated from protobuf enum value: EnableKeepinventory = 3;
     */
    EnableKeepinventory = 3,
    /**
     * @generated from protobuf enum value: KillAllEnderDragons = 4;
     */
    KillAllEnderDragons = 4
}
// ...
class PlayersRequest$Type extends MessageType<PlayersRequest> {
    constructor() {
        super("PayToWin.PlayersRequest", []);
    }
    create(value?: PartialMessage<PlayersRequest>): PlayersRequest {
        const message = globalThis.Object.create((this.messagePrototype!));
        if (value !== undefined)
            reflectionMergePartial<PlayersRequest>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PlayersRequest): PlayersRequest {
        return target ?? this.create();
    }
    internalBinaryWrite(message: PlayersRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message PayToWin.PlayersRequest
 */
export const PlayersRequest = new PlayersRequest$Type();
// ...
class PotionEffect$Type extends MessageType<PotionEffect> {
    constructor() {
        super("PayToWin.PotionEffect", [
            { no: 1, name: "name", kind: "enum", T: () => ["PayToWin.PotionName", PotionName] },
            { no: 2, name: "duration", kind: "scalar", T: 13 /*ScalarType.UINT32*/ },
            { no: 3, name: "amplifier", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<PotionEffect>): PotionEffect {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = 0;
        message.duration = 0;
        message.amplifier = 0;
        if (value !== undefined)
            reflectionMergePartial<PotionEffect>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PotionEffect): PotionEffect {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* PayToWin.PotionName name */ 1:
                    message.name = reader.int32();
                    break;
                case /* uint32 duration */ 2:
                    message.duration = reader.uint32();
                    break;
                case /* uint32 amplifier */ 3:
                    message.amplifier = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PotionEffect, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* PayToWin.PotionName name = 1; */
        if (message.name !== 0)
            writer.tag(1, WireType.Varint).int32(message.name);
        /* uint32 duration = 2; */
        if (message.duration !== 0)
            writer.tag(2, WireType.Varint).uint32(message.duration);
        /* uint32 amplifier = 3; */
        if (message.amplifier !== 0)
            writer.tag(3, WireType.Varint).uint32(message.amplifier);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message PayToWin.PotionEffect
 */
export const PotionEffect = new PotionEffect$Type();
// ...
export const Item = new Item$Type();
// ...
export const SpawnEntity = new SpawnEntity$Type();
// @generated message type with reflection information, may provide speed optimized methods
class EffectRequest$Type extends MessageType<EffectRequest> {
// ...
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: EffectRequest): EffectRequest {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* PayToWin.PotionEffect potion */ 2:
                    message.effect = {
                        oneofKind: "potion",
                        potion: PotionEffect.internalBinaryRead(reader, reader.uint32(), options, (message.effect as any).potion)
                    };
// ...
            }
        }
        return message;
    }
// ...
}
/**
 * @generated MessageType for protobuf message PayToWin.EffectRequest
 */
export const EffectRequest = new EffectRequest$Type();
// ...
export const PayToWin = new ServiceType("PayToWin.PayToWin", [
    { name: "GetPlayers", serverStreaming: true, options: {}, I: PlayersRequest, O: PlayersResponse },
    { name: "ApplyEffect", options: {}, I: EffectRequest, O: EffectResponse }
]);